requirements.txt
alembic==1.4.3
attrs==19.3.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bleach==3.1.4
certifi==2018.1.18
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
entrypoints==0.3
flake8==3.8.4
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-BS4==4.5.3.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.16.0
graphviz==0.14
idna==2.6
imageio==2.8.0
importlib-metadata==3.4.0
importlib-resources==5.1.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipythonblocks==1.9.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jwt==1.1.0
kiwisolver==1.2.0
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
matplotlib==3.2.1
mccabe==0.6.1
mglearn==0.1.7
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.5
notebook==6.0.3
numpy==1.18.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
pika==0.11.2
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
pyaes==1.6.1
pyasn1==0.4.7
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-telegram-bot==10.0.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.18.4
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.10
Telethon==1.10.10
telethon-session-sqlalchemy==0.2.15
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
tzdata==2021.1
urllib3==1.22
visitor==0.1.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
wxPython==4.1.0
zipp==3.4.0
